# fcc-api/laravel/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "FCC API",
        "description": "This is a online API Documentation for FCC",
        "contact": {
            "name": "CJ"
        },
        "version": "1.0.0"
    },
    "host": "fcc-api.wodebox.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/admin/staffs/table/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Assigned tables to a staff",
                "description": "",
                "operationId": "api.admin.assignedtables",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Staff ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "properties": {
                                "tables": {
                                    "type": "string",
                                    "default": "",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Server Error"
                    },
                    "400": {
                        "description": "Logic Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Assigned tables to a staff",
                "description": "",
                "operationId": "api.admin.assignedtables",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Staff ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Server Error"
                    },
                    "400": {
                        "description": "Logic Error"
                    }
                }
            }
        }
    },
...
}